requirements.txt
minio
neo4j
langchain
langchain-community
pypdf
huggingface_hub
python-dotenv
pyyaml
sentence-transformers
transformers
torch
accelerate
bitsandbytes
chonkie
streamlit
qdrant-client
fastapi
uvicorn
pydantic
requests
